software: create liblitescard and move sdcard init/test code to it.
