+++ litex/soc/software/bios/commands/cmd_litesdcard.c
@@ -5,9 +5,10 @@
 
 #include <generated/csr.h>
 
+#include "sdcard.h"
+
 #include "../command.h"
 #include "../helpers.h"
-#include "../sdcard.h"
 
 /**
  * Command "sdclk"
